feat: Refactor existing code to move events under walrus-service
