+++ crates/walrus-service/src/common/utils.rs
@@ -9,7 +9,6 @@ use std::{
     fmt::Debug,
     future::Future,
     net::{IpAddr, Ipv4Addr, SocketAddr},
-    num::Saturating,
     path::{Path, PathBuf},
     pin::Pin,
     str::FromStr,
@@ -22,7 +21,6 @@ use anyhow::{anyhow, Context as _, Result};
 use futures::future::FusedFuture;
 use pin_project::pin_project;
 use prometheus::{HistogramVec, Registry};
-use rand::{rngs::StdRng, Rng, SeedableRng};
 use serde::{
     de::{DeserializeOwned, Error},
     Deserialize,
@@ -88,164 +86,6 @@ pub trait LoadConfig: DeserializeOwned {
             .with_context(|| format!("Unable to load config from {}", path.display()))?;
 
         Ok(serde_yaml::from_reader(reader)?)
-    }
-}
-
-/// The representation of a backoff strategy.
-pub(crate) trait BackoffStrategy {
-    /// Steps the backoff iterator, returning the next delay and advances the backoff.
-    ///
-    /// Returns `None` if the strategy mandates that the consumer should stop backing off.
-    fn next_delay(&mut self) -> Option<Duration>;
-}
-
-#[derive(Debug)]
-pub(crate) struct ExponentialBackoffState {
-    min_backoff: Duration,
-    max_backoff: Duration,
-    sequence_index: u32,
-    max_retries: Option<u32>,
-}
-
-impl ExponentialBackoffState {
-    /// Creates new state with the provided minimum and maximum bounds.
-    ///
-    /// If `max_retries` is `None`, this backoff strategy will keep retrying indefinitely.
-    pub fn new(min_backoff: Duration, max_backoff: Duration, max_retries: Option<u32>) -> Self {
-        Self {
-            min_backoff,
-            max_backoff,
-            sequence_index: 0,
-            max_retries,
-        }
-    }
-
-    /// Creates a new `ExponentialBackoffTracker` that yields an infinite sequence of backoffs
-    /// between the min and max specified.
-    pub fn new_infinite(min_backoff: Duration, max_backoff: Duration) -> Self {
-        Self::new(min_backoff, max_backoff, None)
-    }
-
-    pub fn next_delay<R: Rng + ?Sized>(&mut self, rng: &mut R) -> Option<Duration> {
-        if let Some(max_retries) = self.max_retries {
-            if self.sequence_index >= max_retries {
-                return None;
-            }
-        }
-
-        let next_delay_value = self
-            .min_backoff
-            .saturating_mul(Saturating(2u32).pow(self.sequence_index).0)
-            .saturating_add(Self::random_offset(rng))
-            .min(self.max_backoff);
-
-        self.sequence_index = self.sequence_index.saturating_add(1);
-
-        Some(next_delay_value)
-    }
-
-    fn random_offset<R: Rng + ?Sized>(rng: &mut R) -> Duration {
-        let millis = rng.gen_range(0..=ExponentialBackoff::MAX_RAND_OFFSET_MS);
-        Duration::from_millis(millis)
-    }
-}
-
-/// An iterator over exponential wait durations.
-///
-/// Returns the wait duration for an exponential backoff with a multiplicative factor of 2, and
-/// where each duration includes a random positive offset.
-///
-/// For the `i`-th iterator element and bounds `min_backoff` and `max_backoff`, this returns the
-/// sequence `min(max_backoff, 2^i * min_backoff + rand_i)`.
-#[derive(Debug)]
-pub(crate) struct ExponentialBackoff<R> {
-    state: ExponentialBackoffState,
-    rng: R,
-}
-
-impl ExponentialBackoff<StdRng> {
-    /// Maximum number of milliseconds to randomly add to the delay time.
-    const MAX_RAND_OFFSET_MS: u64 = 1000;
-
-    /// Creates a new iterator with the provided minimum and maximum bounds,
-    /// and seeded with the provided value.
-    ///
-    /// If `max_retries` is `None`, this backoff strategy will keep retrying indefinitely.
-    pub fn new_with_seed(
-        min_backoff: Duration,
-        max_backoff: Duration,
-        max_retries: Option<u32>,
-        seed: u64,
-    ) -> ExponentialBackoff<StdRng> {
-        ExponentialBackoff::<StdRng>::new_with_rng(
-            min_backoff,
-            max_backoff,
-            max_retries,
-            StdRng::seed_from_u64(seed),
-        )
-    }
-}
-
-impl<R: Rng> ExponentialBackoff<R> {
-    /// Creates a new iterator with the provided minimum and maximum bounds, with the provided
-    /// iterator.
-    ///
-    /// If `max_retries` is `None`, this backoff strategy will keep retrying indefinitely.
-    pub fn new_with_rng(
-        min_backoff: Duration,
-        max_backoff: Duration,
-        max_retries: Option<u32>,
-        rng: R,
-    ) -> Self {
-        Self {
-            state: ExponentialBackoffState::new(min_backoff, max_backoff, max_retries),
-            rng,
-        }
-    }
-
-    fn next_delay(&mut self) -> Option<Duration> {
-        self.state.next_delay(&mut self.rng)
-    }
-}
-
-impl<R: Rng> Iterator for ExponentialBackoff<R> {
-    type Item = Duration;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.next_delay()
-    }
-}
-
-impl<I> BackoffStrategy for I
-where
-    I: Iterator<Item = Duration>,
-{
-    fn next_delay(&mut self) -> Option<Duration> {
-        self.next()
-    }
-}
-
-/// Trait to unify checking for success on `Result` and `Option` types.
-pub(crate) trait SuccessOrFailure {
-    /// Returns true iff the value is considered successful.
-    fn is_success(&self) -> bool;
-}
-
-impl<T, E> SuccessOrFailure for Result<T, E> {
-    fn is_success(&self) -> bool {
-        self.is_ok()
-    }
-}
-
-impl<T> SuccessOrFailure for Option<T> {
-    fn is_success(&self) -> bool {
-        self.is_some()
-    }
-}
-
-impl SuccessOrFailure for bool {
-    fn is_success(&self) -> bool {
-        *self
     }
 }
 
@@ -364,32 +204,6 @@ where
     fn drop(self: Pin<&mut Self>) {
         if !self.is_terminated() {
             self.observe(None);
-        }
-    }
-}
-
-/// Repeatedly calls the provided function with the provided backoff strategy until it returns
-/// successfully.
-pub(crate) async fn retry<S, F, T, Fut>(mut strategy: S, mut func: F) -> T
-where
-    S: BackoffStrategy,
-    F: FnMut() -> Fut,
-    T: SuccessOrFailure,
-    Fut: Future<Output = T>,
-{
-    loop {
-        let value = func().await;
-
-        if value.is_success() {
-            return value;
-        }
-
-        if let Some(delay) = strategy.next_delay() {
-            tracing::debug!(?delay, "attempt failed, waiting before retrying");
-            tokio::time::sleep(delay).await;
-        } else {
-            tracing::debug!("last attempt failed, returning last failure value");
-            return value;
         }
     }
 }
@@ -737,54 +551,6 @@ mod tests {
     use walrus_test_utils::{assert_unordered_eq, param_test};
 
     use super::*;
-
-    mod exponential_backoff {
-        use super::*;
-
-        #[test]
-        fn backoff_is_exponential() {
-            let min = Duration::from_millis(500);
-            let max = Duration::from_secs(3600);
-
-            let expected: Vec<_> = (0u32..)
-                .map(|i| min * 2u32.pow(i))
-                .take_while(|d| *d < max)
-                .chain([max; 2])
-                .collect();
-
-            let actual: Vec<_> = ExponentialBackoff::new_with_seed(min, max, None, 42)
-                .take(expected.len())
-                .collect();
-
-            assert_eq!(expected.len(), actual.len());
-            assert_ne!(expected, actual, "retries must have a random component");
-
-            for (expected, actual) in expected.iter().zip(actual) {
-                let expected_min = *expected;
-                let expected_max =
-                    *expected + Duration::from_millis(ExponentialBackoff::MAX_RAND_OFFSET_MS);
-
-                assert!(actual >= expected_min, "{actual:?} >= {expected_min:?}");
-                assert!(actual <= expected_max);
-            }
-        }
-
-        #[test]
-        fn backoff_stops_after_max_retries() {
-            let retries = 5;
-            let mut strategy = ExponentialBackoff::new_with_seed(
-                Duration::from_millis(1),
-                Duration::from_millis(5),
-                Some(retries),
-                42,
-            );
-            let mut actual = 0;
-            while let Some(_d) = strategy.next_delay() {
-                actual += 1;
-            }
-            assert_eq!(retries, actual);
-        }
-    }
 
     mod byte_count {
         use super::*;
