+++ crates/walrus-service/src/node/storage.rs
@@ -44,9 +44,6 @@ pub use database_config::DatabaseConfig;
 mod event_cursor_table;
 
 mod event_sequencer;
-
-pub mod event_blob;
-pub mod event_blob_writer;
 mod shard;
 
 pub(crate) use shard::{ShardStatus, ShardStorage};
